Add "export source as (inline) styled HTML (body only)"
... and refactor the functions out a little bit

It would be easy to add inline styled export for the whole document,
but I don't think that makes any sense.

When there will be a new dialog interface for this
we won't need the individual entry functions anymore.
Instead the flags will be set according to the Dialog's results