Source-Export: QSettings start
Export options are now stored with the global QSettings object.
They are persistent throughout the session,
can be saved explicitely (from the "Export source" dialog) or
implicitely (Checkbox in Preferences->General)

The Dialog itself is a mere dummy yet. At least it shows
disabling of the "Save" button, Cancel and OK work too
(Cancel clears the clipboard)

The export options can't be edited in this intermediate state.
Default values can be set in export/__init__.py,
but they will be overwritten through the Settings.