auto-complete path of LilyPond binary inside .app bundle on Mac OS X
The LilyPond binary in Mac OS X is located in a subfolder of the
application bundle. From the 'Select a file' dialog the user can only
select the application bundle and not its contents.
With this change the user can select e.g. '/Applications/LilyPond.app'
and have it automatically completed to the corresponding full path
'/Applications/LilyPond.app/Contents/Resources/bin/lilypond'.