preview-mode: Stub for configuration UI
Configuration variables can be inserted in the document
through the -e commandline option.

This commit shows how to do it and hard-codes anchor-point dots green.

TODO: Complete the function that reads out all values
from the settings.
Try to make sure that only values are added that don't
match the default settings (in order to avoid too long
command lines.

Start Preferences page
-----------------------------
This is really a WIP, just started to show where work has to be done.
Add a new page to the preferences
(which is currently a mere copy of the LilyPond preferences),
add a dummy help page.

Hide work on Preferences UI
------------------------------------
The work on the UI for the configuration Preferences
is postponed to a future release.
Instead of completely reverting the stubs already done
this commit simply hides them so they aren't executed and displayed