Add git module (basic implementation)
Add a module with a Git class to manage Git repositories.
This commit implements an initial set of functions.

If Frescobaldi is running from git the module is imported and
an object git.app is created representing the application's
Git repository.

Later we will also use it to manage documents' Git repositories.