Add Git menu
For now simply list all local branches.
Clicking on a branch tries to checkout this and restart the application.
If git gives an error (e.g. non-clean working tree) it aborts with a message box.
If checkout succeeds the user gets a message box and can restart
Frescobaldi now or later.