preview-mode: Improve coloring of voices
Now also manual stems benefit from the >=2.17.6 improvement.
They use \temporary\override then, so after reverting the
stem direction to neutral it won't revert to black but to the
previously active color.

As the other colorings (slurs, dynamics) don't interfere
with the coloring of voices I consider this issue fixed
with this commit.