Mac OS X: fix application bundle detection to set MenuRoles
Using py2app, sys.argv[0] is the original script instead of the
executable inside the application bundle.
To check if the script is executed from the application bundle, py2app
sets sys.frozen to "macosx_app".
The sys.argv[0] check is kept to maintain compatibility with other ways
of creating the application bundle (cx_Freeze, Platypus?).