svgedit: handle group dragging WIP
This is a start, but I don't get to the end with it now.

The actual setTranslate is moved to the DraggableObject itself,
which I think is good.

What I intended is to give this object a group property if it is
part of a group (in mousedown), but I don't see how to achieve that
(I thought I could compare to the draggable array, but didn't manage to.