Fix exporting MusicXML in Python 3
ElementTree.write() should get encoding="unicode" for files opened in
text mode, but it would ignore the encoding argument to MusicXML.write().

To respect the encoding argument, open the output file in binary mode.
Substitute the encoding in the XML declaration, don't blindly write
'UTF-8' is it's not. Close the output file when done.