Implementing rest tools from ly.rests as Frescobaldi tools
Replace full measure rests with spacer rests. Change all R to s in the document or in a selection.

Replace spacer rests with full measure rests. Change all s to R in the document or in a selection

Replace positioned rests with plain rests. Change all \rest with r in the document or in a selection.

Also see issue #497.