Check if ly is installed and has the correct version.
If python-ly is not available or too old, we display a critical message
and exit.  In the future we will probably remove this function.
In a good software distribution this should never happen, but it can happen
when a user leaves the old stale 'ly' package in the frescobaldi_app
directory, or has not installed python-ly at all.

Because that yield unexpected behaviour and error messages, we better
check for the availability of the 'ly' module beforehand.
Importing ly.pkginfo is not expensive.