really fix #650
the disconnected signal could occur before all commands were handled,
causing a crash due to the socket being garbage collected. This
occurred mainly when an event loop was running during display of the
error dialog when a non-existing file was requested.

now we dont connect to the "disconnected" signal anymore, but check
if the socket is still connected after reading all data.