Make manuscript viewer open single file (pdf) only
For now the manuscript viewer can only handle a single PDF
(and I think it's a good idea to leave it at that for now).
Therefore the the file open dialog should only allow opening one file,
and as we can only open PDFs for now this should also be restricted.

Once we open up the tool it'll be easy to re-add the functionality.