Move AbstractViewerPanel in its own module (WIP)
This is the next step towards having a separate abstract viewer class.
For now it seems to be (knock-on-wood) without any change in behaviour.
The abstract class was moved to the 'viewers' module.
The images.py and documents.py files were also moved there.
The other files are likely to be specific to the music view, but might
be moved later.

For now the Widget has been left in the musicview directory. There is
presumably much common functionality in here too that should be moved
later, but I wanted to have a manageable small step.

With regard to functionality it has to be seen if we will need another
intermediate class to switch between the SVG and the other viewers (i.e.
between the qtpopplerview and the QtWebView).
But that can be left for later.