Make musicview a submodule of viewers
Move the whole musicview module *inside* viewers.
If we have an abstract "Viewers" class and several concrete viewers that
inherit from it it will be cleaner to reflect that in the file structure
too (also making the main directory list slightly shorter).

We may well intend to add further viewers in the future, e.g. a Verovio
viewer or an annotation viewer/editor or whatever.