manuscript: start file open at currently opened manuscript dir
When a manuscript is currently opened in the viewer its directory
will supersede the directory of the currently opened editor document.

Addresses #711.

However, I'm not completely sure if that "smart" behaviour might actually
be perceived as inconsistent by the user. Do we need to make that
configurable in the Preferences?