Make AbstractPopplerView always contain the toolbar
The implementation of optionally including a toolbar if 'use_layout'
is present was somewhat quirky. And actually we *will* want to always
have a toolbar present.

TODO: The generation of the toolbar itself should be done in a similar
modular way as the context menu so common things can be done in the base
class while specific things can be overridden.