Viewer: Fix positioning upon session change on app startup
As described in #735 the restoration of the current document's position
was broken. *With* kinetic scrolling the offset was wrong, *without* it
didn't work at all.

The kinetic scrolling issue turned out to be a deliberate limitation by
the designer of the kinetic module, but we ignore that because we suppress
kinetic scrolling for session change.

The issue was that a resize event is triggered upon application startup,
and this happens *after* setting the position. And in this resizing event
the position at the top of the file is given. The kinetic scrolling did
actually hide this fact.