Add AbstractViewWidget base class (stub)
This is setting up a new switch in the class hierarchy.
At some point all viewers should be included in the common hierarchy in
order to provide a consistent interface and reuse as much code as
possible.

Right now we only have the poppler widgets in mind, but we will also want
to support the QWebView widgets in the same context.

For now this is an empty stub but we will move interface and large parts
of the implementation from AbstractPopplerWidget here later.