Avoiding the position to get reset by following resize events
This is a temporary hack
(that it breaks the convention of protected variables is not the only problem, or differently put, there's a reason why the convention is used).

If f29d74bd26a21124a4f6d1164967a93d8752b7bf gets accepted this can be reverted. If not a better workaround should still be found.

The issue this addresses is described in #735.