One more try to properly save and restore manuscript session
This is the first commit and should be followed by a second.
This still adds the documents one by one, but the active manuscript
is not resorted to the end of the list.
Also we don't handle the position of the current file independently, but
treat the active manuscript as one with the others.