manuscript: use reworked loadManuscripts()
loadManuscripts() now doesn't sort the manuscripts anymore. Sorting
is deferred to a - still non-functional - separate method (code is only
stuffed away in an unused method for now).

The method handles a list of manuscripts that can be filenames or
tuples of filename and position. (These can even be individually, as the
check is done per entry).
if an active manuscript is given (well, this *should* always be there,
but we check anyway) this is respected, but not through adding it to
the list at the end but keeping it at its position.

The session change slot now uses this method and could therefore be
drastically simplified.