Move session handling from manuscript widget to viewer panel
This is a generic task that  would be expected to be handled by the panel
and not by the widget.
With the previous commit it should be possible to have new viewers (e.g.
the reimplemented musicview) to simply inherit and have all the session
handling in place automatically. (The music view will want to override
the reportMissingDocuments method with a 'pass' stub, though.