Suppress calling of handler for empty session
Explicitly calling the session handler upon creation of the widget
had been implemented for the case that the viewer tool is closed on
application startup. In that case the session has to be loaded explicitly.

This commit filters out that call when there is no current session to
load.

Relates to #760