Implement viewer's help button as a regular action
When trying to deal with #760 it became clear that it doesn't really
make sense to explicitly deal with the help button as a QToolButton
and add it as a widget to the toolbar.
It's much more straightforward to manage it as a normal QAction, just
like everything else. Besides, it became apparent that the context menu
did also manually create its own instance of the thing.

NB: This doesn't change anything with #760 itself, but is a coding
improvement anyway.