Factor out viewers' toolbars
Moving the toolbars to their own module makes sense because they have
got some significant weight by now. For example it was an indicator that this refactoring is a good idea that I could rename a number of _tbAddSomething() methods to addSomething() etc.

This opens up the way for a custom QToolBar implementation