Allow MusicXML export to follow \include's
This is functionality that has already been available with python-ly
but simply not "activated" in Frescobaldi.

Closes #873

I hope this is sufficient as it simply passes the filename of the
currently open tab. If anything like respecting master documents
should be respected some more work would be involved (but actually
I don't think that should prevent this from being merged).