Add library path to LilyPond invocation (WIP)
Not adding this can cause compilation failures when the bundled
versions of libraries (like Ghostscript) don't match the system-provided
ones.

Note that this commit only is checked with downloaded and self-compiled
versions on Linux. I assume distro packages will work as well (or
don't need it anyway), but I have no idea about Mac and Win.