Fix autocompleter when overwriting selection (sorry)
Partial matches are directly inserted in the document as selected,
and typing the next letter accepts the selection by continuing
afterwards.

However, this turns out to be a problem when continuing to enter
the original word. So this commit fixes the use case of typing a letter
that happens to be the first letter of the selection. In that case this letter
will simply be overwritten and the selection starts one character later.