Another Fix (another sorry)
Working more with the new feature revealed that Janek's suggestion
of automatically accepting the partial completions isn't good at all.
If I want to write "author" and upon "au" I get "auto..." as suggestions
then continuing to type will seriously mess up things.

So now (again), when a partial completion is present as a selection
*any* new text will remove the selection and add the character *instead*
of the selection.
Interestingly the behaviour is nearly the same when manually typing
the suggested characters.