Goto line: try to place the cursor at the same character position
As the popup dialog appears at the cursor position it is clearer for the
eye to follow if the cursor will be placed at that position too
(instead of at the beginning of the line).
If the target line is too short the cursor is placed at the end of the line.

(No precaution has been taken (yet) for broken lines).