Fix "landscape" handling in score wizard (effective LilyPond 2.19.60)
Addresses #768

Frescobaldi used a wrong syntax for specifying landscape paper
in the score wizard. The additional 'landscape argument to set-paper-size
is actually meant to rotate the print on the page and not the page itself.
From LilyPond 2.19.60 this is properly handled as intended, so Frescobaldi
*actually* produces wrong output.

This commit produces correct output - and this will actually work
backwards at least to 2.18.