Add "First system only" option to Engrave (Custom)
Addresses #724

The "Engraving Mode" combo box in the Engrave (Custom) dialog
gets an entry "First System Only", which adds the
-dpreview and
-dno-print-pages options
options to the command line.

This makes sure that only the first system is compiled, as PDF and PNG.

One remaining problem is that now I'd like to view that preview in the
music view, isn't it?