Use include targets retrieved from mouseMove to open includes
As we have already determined the list of included files during the
mouseMove event we can simply use that for opening the files
instead of running throught the filenames_at_cursor procedure again.

Relates to #947, I would like to narrow it down to the single string
under the cursor instead of the whole block.