Remove obsolete open_file_at_cursor.genToolTipInfo
As discussed in https://github.com/wbsoft/frescobaldi/pull/941#issuecomment-302423308 this function has been
superseded by other functionality.

Essentially it was an issue with performance with large documents
that always the whole document structure had to be parsed.
In the new approach only the line under the mouse pointer
is considered.