Make vcs interface more generic
All repo classes have the same name Repo() now and are addressed
using the module name exclusively. This makes it possible to use
module names from variables and have a transparent interface
regardless of the type of the actual repository.

Each Repo class must implement a vcs_available() class method that
returns True if the associated VCS is installed.
_run_git_command() function has been moved to a class method as well,
but that's at the discretion of each Repo class.