Some minor questions
I've added my questions as comments so you can see them easily
in this commit @wenxin-bupt. And I've added them to a separate branch that we can simply delete later.

I have read through all the files in the vcs module, except the diff_hunk and the diff_lines functions, and apart from the fact that it's still not integrated in a running version (so I only read it an didn't test anything) everything looks very fine.

You have found a way to merge the signal-slot concept with your approach of callback functions, which is probably not completely in line with the Qt way of things, but I don't see any problems with it anymore :-)

So if you give me some hints on how to set up a way to test the code in a running Frescobaldi I'll have another look including the core diff functions.