vcs: WARNING about cause of crash!
The function originally intended to check the presence of Git
is never checked.
And Frescobaldi crashes when trying to open a file from a Git
repository when Git is not installed.
I'm not sure if this is actually because the function isn't called
or if there's a bug in another function. In either case, this is
critical.