VCS: Move code from Git Document to Abstract Document
Much of the code that has been added to the Git document is
pretty generic.
This commit directly lifts a number of methods and fields to the base
class and singles out steps through the template pattern in other cases.

There are a number of (large) methods that have been left in the Gitdoc
class because I'm not sure about the implications for other VCSes.
Probably we'll have to refactor these functions *if* one day other
VCSes should be supported as well.