Remove vcs.helper.py
The coding in this class suffered from too little abstraction,
it was not necessary to copy the classes for all VCS types.
Together with the changes in the previous commits it became
apparent that it makes sense to completely avoid this module.

In general we're approaching a state where we don't explicitly
invoke classes or functions for specific VCS types but work on
a generic API, against objects whose classes have been looked
up from a dictionary.