#1028 (VCS): Streamline VCS error reporting
Use one single method in vcs.VCS to report VCS errors.
Don't print to the console but show a Message Box instead.

NOTE: This required further fixes to an existing disable() method
because when trying to disabling a file more work had to be done
with the connections and references.