VCS: Prepare "Version" menu to replace "Git"
So far the "Git" menu was used when Frescobaldi is run from Git.
However, this is actually the place to add VCS actions.

This commit renames the menu and makes it visible whenever
VCS is available at all.
When Frescobaldi is run from Git the previous Git menu is shown as
a submenut now.

#940