#866
Not sure how to optimize the regular expression to fit lilypond note names exclusively. Now every 3 or less letters word is a note name if it doesn't start with #, \, _, -, ^ symbols and doesn't end with dot. Of course it also finds such 'notes' within quotes if happen. I am not experienced in Python nor in regular expressions so I give up to make it better for now. From the other side it's pretty usable even so.