WIP: Initial dialog for showing available fonts
This is the first installment fo an improved dialog for showing available
fonts. So far this parses the results to a number of dictionaries and lists
but doesn't do anything with it.

TODO: Properly parse the details of a font family
TODO: Create a new custom widget:
  - show all font families as a table
  - each table row can be expanded to show more details about the family
TODO: Search/Filter
  - An additional text input allows for incremental filtering for font families
QUESTION: Do we need to show config files and font dirs?