Use TabWidget to display both the raw LilyPond output and the results
Displaying the log output lets the user inspect the raw material if they
want to. In addition it makes the waiting time for the results feel more
"natural".
The results are displayed in the second tab which is automatically
opened when the job has completed.

TODO:
So far this still uses a Log() for the result display, which is completely
inappropriate.