Properly populate the (preliminary) data model for font display
This commit fixes an issue with parsing and attaching multiple styles.
Additionally it fleshes out the routine to display the fonts in hierarchical
order. So far this is done into the old Log widget but the logic behind
the "tree building" should be portable to any new approach using the
model-view structure.