Create "Tools=>Format Code" submenu
I think the Tools menu is too long and should be restructured.
From that perspective it would be good to also move the section with
Pitch, Rhythm etc. to a submenu, but that would probably make it
too complicated to access.

The current submenu for code formatting moves a number of items
out of sight, and especially I intend to add more commands in the
future, like some "linting" functions such as forcing/removing whitespace
between elements or forcing (custom) order of attached objects (dynamics,
articulation, markup etc.)