Separate "available fonts" storage and dialog
There is now a textfonts.available_fonts object that will live throughout
the application. It is responsible for loading and maintaining the list of
available fonts.

The "show available fonts" dialog has been factored out in its own file
and strictly *displays* the data from the main storage.

This will make it possible for *other* widgets in the future to make use
of the available fonts list.

TODO:
Currently there is still some parsing and matching done in the dialog's
populate_tree function. This logic should also be moved to the initial
parsing process in the main TextFonts() object.