Better separation of concerns
Move the (problematic) heuristic of matching styles to font subfamilies
from the dialog to the TextFonts class.
It is important to store the complete info about available fonts there
and as little as possible in the dialog to make the data available to
other UI elements as well (e.g. the "Document Fonts" dialog) and others,
maybe a small text editor component in the future.