Move MusicPreviewJob (musicpreview.py) to job.lilypond.VolatileTextJob
This type of job (document is passed in as a text string) seems to be
more generically usable, so it shouldn't be defined in the special
dialog's domain.
This is a job for preview or other types of compilations where the
input is not a document in a file but a simple string. A temporary
file is created and set as the job's document. No point-and-click links
are created (since they don't have any target to point to).