Refactor Job() code completely
This is a thorough reorganization of the code, and I hope it goes
in the direction of better extensibility. Generally it reduces the need
to redo the same job configuration in many different places and
gives more responsibilities to the different LilyPond job classes.

This should be thoroughly tested!

NOTE: This breaks the "Show available fonts" function, but I would
rather not touch it because that will be handled in the separate PR
#1075 anyway.