Remove job.queue.Scheduler, replace with JobQueue (WIP)
The Scheduler class (originally process.Scheduler) was a light-weight
version of the job queue and is replaced with a single-runner
job queue (to avoid having to maintain the same functionality in
two classes).

The WIP refers to the fact that here we instantiate a separate
JobQueue, but we should actually integrate it in the global
job queue infrastructure once that's ready.