Initial set-up of app.job_queue()
This instantiates one global JobQueue in continuous mode, for now with two runners.
It is immediately possible to add jobs to that queue, which for now will be
done for the jobs in lilypondinfo.py.

TODO:
1)
Store the number of runners in the Preferences (ensure proper
behaviour when changing that number)
2)
Assign runners to different "task-groups", e.g.
- 1 runner for internal tasks (=> crawl for autocomplete data etc.)
- all-1 for LilyPond compilations