Various improvements to JobQueue and Runner
Runners now have their index stored as a field, and when starting
a job the runner adds a job._runner reference to itself.

Fix the logic of JobQueue.add_job().
Originally it hadn't been properly handled to decide when to reject
adding a job and when to implicitly start the new job (we shouldn't
only start the job immediately when the queue is idle but when
*any* runner is idle).

Incidentally fix issues with not having super.__init__ called in some
classes.

Adding comprehensive signals for the JobQueue
Review coding and comments