#1066: Respect Logtool font size setting (#1115)
Closes #1066
There seems to be a conflict between letting the log font be zoomed
and respecting the Preference for font size.

So this makes the Log Tool consistent with how font sizes are managed
in the document editor. But the ability to *zoom* the log with the
mouse or keyboard shortcuts (?) is removed by that commit.