Require Python 3.3
I stumbled over main.patch_pyqt and think this should be dropped.
Essentially this was necessary for Python2 and should *not* be necessary
for Python3 anymore.

Based on the code in the function this "patch" was necessary until Python 3.2,
but given the fact that Python 3.3 was released in 2013 and has reached
end-of-life I think it's safe to drop it altogether and bump the requirement
to Python 3.3.

This is also related to #1126.