Implicitly create an extension's Tool Panel
It turns out that it is not necessary for an extension to explicitly
implement a subclass of ExtensionPanel. There are only two properties
that are actually custom (the widget class and the initial dock area),
therefore this commit gives an extension the option to define these
properties. If this has been done then a Panel will transparently be
created.