Adjust VBCL to Frescobaldi's needs
Originally the VBCL module had been designed specifically for openLilyLib
packages. In Frescobaldi it has to become somewhat more generic in order
to become useful for future generic use. For now we need it to handle
extension metadata, but it will also be available for other extensions
(for example openLilyLib, again).

Therefore, the data against to check must not be built into the module
but passed from outside (here we'll do it from the extensions package).