Separate loading of extension info and code
This commit separates the loading of extension info and code.
This is mainly because the extension info has to be loaded regardless
of the 'active' state of extensions in general or individual extensions.

Only subdirectories containing 'extension.cnf' are now considered extensions.

If there are errors either parsing the info or loading the extension a
summary message box is created.
Information about failed extensions is displayed in the Preferences page