Support extension dependencies, improve reporting
The "dependencies" entry in extension.cnf is now respected.
check_dependencies() detects:
- missing dependencies
- deactivated dependencies
- circular dependencies

In any of these cases the warning message is displayed, and extensions
are not loaded at all (*maybe* one can improve that in the future).
Otherwise the loading order of extensions is guaranteed to respect
dependencies (so extensions an extension depends on are loaded earlier).