Settings track modified state of individual settings
This is needed to allow extensions to respond to settings changes.
Since now settings are stored centrally the extension doesn't know
if a setting has changed. Instead of maintaining a copy of the settings
solely for this purpose the extension can now ask the settings object
for the modified state of a setting and reset it after responding to it.