#1088 Autocompleter: Don't treat dot like ENTER
#1088 describes how the fact that entering a dot makes editing harder
in cases when there is no completion intended (`\tweak Beam.` should
*not* complete to `beam-thickness`)

This commit removes the link from Enter/Return to `.`. That means that
entering the dot will *not* automatically complete to the first match
in the completer popup. If that is wanted one would now have to enter
ENTER plus the dot. This need for one more keystroke seems more than
worth it, considering less unintended completions.