CachedPreviewJob class
This is a sibling to VolatileTextJob, a LilyPondJob whose
input is a string instead of a file. But other than the
volatile text job the cached preview stores the file in a cache
and reloads the PDF instead of recompiling the file if it is found.
For persistent caching a specific cache directory can be passed
to the __init__ method, otherwise a temporary file will be used.

The music preview conditionally makes use of the new job type when
it is passed cached=True.