Reimplement the font command generation
Completely overhauled the generation of font commands.
Allow generic LilyPond and openLilyLib styles.
"Full" and "regular" commands (full => use in preview or wizards)
Handle triggering of sample compilation.

Dialog now provides the properties
- font_cmd()
- font_full_cmd()
that can be read after closing the dialog.

NOTE: Text font families are handled by the command generation,
but the names are still hardcoded to be the LilyPond defaults.