Heuristic to filter notation fonts from text font list
While it isn't possible to detect a font as "notation" font
this commit compares the font name with the list of notation
fonts that are installed in the current LilyPond.
So they are suppressed in the tree view. (However, this fails
(by still displaying the font) when the font is installed as
system font but not in the current LilyPond.