(Automatically) Install fonts from repository
The "Music font repository" is a local directory where the
user can save their notation fonts. If that is set all fonts
can be directly installed without having to look up the directory
with the "Install..." button.

With the "Auto install" option this will be done whenever the
font dialog is opened, so any new LilyPond installation can
automatically be provided with the notation fonts.