Integrate new userguide page for Document Fonts (WIP)
This doesn't properly work yet and is only committed
in order to make it possible to test the issue.
As explained in a comment to #1169 the user guide is
displayed *behind* the dialog (or even the application window),
or probably the other way round: the dialog insists on
remaining at the front.