UrlRequester: Several improvements
This replaces Pull Request #949 with proper implementations
(and more)

# Add fileMode argument to __init__

While not changing the default this allows setting the
filemode upon instantiation.

# Add mustExist validagion

If set to True the file must exist. As long as the path set
in the lineEdit doesn't exist the text is colored,
and when trying to leave the widget while it points to a
non-existent path it is instead reset to its previous value.

# Add editingFinished signal

Signal is emitted when the editing has finished.
Treats the lineEdit and the dialog as one entity:
When the lineEdit loses focus because the button
has been clicked the signal is *not* emitted.