CachedPreviewJob and more coding in job module
A Cached PreviewJob based on a volatile text job,
that is aware of available compilations
(based on content hashing of the text document).
The full text is hashed but includes are not flattened
(maybe something interesting, see lyluatex for an example).

By default caching is done in the temp directory and discarded
upon program termination, but if a persistent writable path
is provided caching can also be done persistently.