fix waiting overlay not centered when window is resized
it is not needed to call resize() inside resizeEvent(); resizeEvent() is
called after resize(). We want to adapt our size to the parent, a simple
way to do that is just on start() adapt ourselves to the parent size.

If the parent is resized *during* display of the overlay, our size is not
adapted. To do that, we could use an eventFilter construction and get all
the resize events of the parent and adapt outselves. But that's a bit
overkill I guess, as the waiting overlay doesn't last long.