Complete review of job.queue
This hopefully clears many things up with regard to job queues.
Responsibilities have better been separated between
Job, Runner, JobQueue and GlobalJobQueue.
Specifically we don't iterate over runners and check their state,
instead a runner will be referenced in a _running or a
_shared_running list, and the state of these lists determines
the state of the queue (for example: when both lists are empty
the queue is idle).

Handling of shared and dedicated runners seems to be done
properly now, but no efforts have been takien whatsoever to
manage the distribution of resources (i.e. validity checks
when runners are requested in the Preferences).