neater way to clear the page-set setting that could erroneously have been left behind
Although QPrintDialog does not read back the cups properties, the other
options are correcly set when the user selects Ok. Only a previous page-set
setting was not removed even if the user selects All Pages in the dialog.
So we remove the setting beforehand, and if PageSet even or odd is/was
selected in the dialog, it is aded again correctly.