The real fix for the previous commit
I can't explain to myself why this *seemed* to work.
After a few branch checkouts it doesn't do so anymore.
And the reason is pretty obvious: There is no `QStringList`
in the *Python* implementtion of Qt5 anymore (because there
is no `QString` anymore either).

This creates some hickups in the Python implementation of
`QSettings` and requires a workaround that had long been
implemented in
dd5e2e9b85faefa4187ff7e7d108d4b8610bb830
and is pretty widely used across the application.

So this should fix this one too.