fix printing on Mac
Now that printing with the Arthur backend seems to work correctly,
there is no need to direct the user to open the PDF file with an
external program.  Instead, when printing directly to CUPS, the user
is warned about its drawbacks.  Accordingly, the default value of the
printing/directcups preference is set to False on Mac.

Fixes #30