Fix compiling source files with non-ASCII characters with LilyPond 2.23.6+
In "Run LilyPond with English messages" mode, Frescobaldi used to set
the LC_ALL and LANG environment variable for the LilyPond process.
With versions of LilyPond that use Guile 2 (2.23.7 and later, 2.23.6
for some, also 2.22 from Homebrew), this breaks compiling files with
non-ASCII characters because unlike Guile 1, Guile 2 actually decodes
the filename string from the command line, and does so based on the
locale encoding.  Fix this by only setting LC_MESSAGES, which is
enough to stop localization of log messages without messing with the
encoding used to decode command line arguments.