Fix translateUI() memory management
Before, the translateUI() function connected a signal to its parameter's
translateUI() method, and never disconnected it. This had two
undesirable consequences: first, this would completely prevent garbage
collection for these objects, and second, it went wrong in case the
underlying C++ object behind a PyQt widget was deleted and the signal
later tried to retranslate it. Instead, translateUI() only stores weak
references now, allowing the retranslatable objects to still be
garbage-collected, and for QObjects, it removes the connection to the
signal as soon as the underlying C++ object is deleted.

Fixes #1505