Don't try to get sip version
This doesn't really make sense, since sip is used for building
python-poppler-qt5 but is not used at runtime.  If the user happens to
have sip installed, it may be a different version than the one used to
build python-poppler-qt5. As far as I understand, the PyQt5.sip module
is generated by sip and does include the version of sip that was used to
build PyQt5, but even that does not tell which sip version was used for
building python-poppler-qt5, which may be a different one.

Fixes #1332