Use "org.frescobaldi.Frescobaldi" as icon name
The file in frescobaldi_app/icons/ is called "org.frescobaldi.Frescobaldi.svg".

This fixes the icon displayed at the window's top left of on Windows.

Note that this is unrelated to the file
frescobaldi_app/icons/frescobaldi.ico, which is used in the Windows
installer but not handled by Qt.