Use standard gettext module (except in i18n/)
Now that Frescobaldi requires Python 3.8, which is the first version of
Python where the gettext module supports the message context feature,
the custom MO file reader is not needed anymore, for Frescobaldi itself.
This replaces it with plain standard gettext (and adapts the
infrastructure to put MO files where gettext expects them).

On the other hand, it's still needed in i18n/ for the molint.py script.
The plan is to obviate it there in a second step, for example by
replacing molint.py with https://pypi.org/project/lint-po .

This partly addresses #1285