Move to tox, replacing i18n/Makefile and linux/Makefile
tox is a popular Python automation tool. It is cross-platform (without
shell-specific peculiarities) and can automatically install Python
dependencies.

The intent of this change is twofold. It will improve the experience for
Windows contributors, for whom it can be difficult to get Make. (A
future extension could be to rely on Python tools instead of Gettext.)
And it will allow getting the right version of Sphinx automatically when
the user guide is converted to Sphinx.

Fixes #1113