i18n: Remove the deduplication between the frescobaldi and userguide domains
The previous po-update script took care of not inserting messages into
userguide PO files if they were already in the frescobaldi PO files, to
avoid duplicate work. However, this complicates the Sphinx support
because we cannot use the existing Sphinx translation workflow tools due
to this peculiarity. Moreover, this might lead to messages that need
context to be translated differently between frescobaldi and userguide,
but it's suboptimal because there is no way to add a translation context
in the userguide.

This adds a copy of all those messages to the userguide PO file and
adapts the PO/MO generation scripts to make the frescobaldi and
userguide domains independent.