Rework point-and-click to not depend on Poppler
Point-and-click now works with the QtPdf backend. This also
required some major changes to qpageview, so make sure you are
running the latest version of the Qt 6 port.

There is currently some visual glitching with the highlight
rectangle drawn after clicking a link, which does not otherwise
affect functionality. I suspect there may be an incorrect unit
conversion happening somewhere in qpageview.